sql: refactor
